remove importlib usage in enum converter
Summary:
The `py3-types` target already directly depends on `python-types` target, so direct import of `.thrift_types` is guaranteed to succeed. Remove `importlib` usage so this works better in cinder environment.

This works for the 99.99% of thrift files that produce valid `.thrift_type` thrift-python files. There is a long tail of `.thrift` that produce valid thrift-py3 but invalid thrift-python, many due to a combination of `namespace py3 ""` + name collision between module name  and struct name.

For this long tail, we have to put the `import` in a `try` statement, then defer the failure to call to `._to_python()`.

Reviewed By: yoney

Differential Revision: D65087762

fbshipit-source-id: 703b1f2a865c14e65479e0ce9a7581ca352cdf09