Fix/improve annotation bundling
Summary:
Add unit tests for bundled annotations. To this end introduce an option to disable adding standard annotation and lib directories to the search paths in compiler tests.

This revealed a bug in the placement of the logic to inject the bundled annotation - fix it by moving the injection before parsing.

Move the warning about missing `schema.thrift` to the schema injection path where it belongs - this warning interferes with testing and is mostly noise anyway unless schema is actually needed.

Reduce conditional compilation scope.

Reviewed By: thedavekwon

Differential Revision: D65163830

fbshipit-source-id: bd42d3f2e7633e3f8b9d282947cd994ed11c2664