Fix transitive includes aggregation for schema bundling
Summary:
{D61240139} had an off by one bug.

If you look at the list of includes...
```
  mstch::node transitive_schema_initializers() {
    const auto& includes = context_.cache().get(
        *program_, [&] { return gen_transitive_include_map(program_); });
    mstch::array initializers = {
        fmt::format("{}()", schematizer::name_schema(sm_, *program_))};
    for (const auto& [_, include] : includes) {
      initializers.push_back(include);
    }
    return initializers;
  }
```
Notice that the first entry is *always* the program itself.

The code affected by this diff flattens the set of transitive includes. For example, Given the following include hierarchy `A.thrift` → `B.thrift` → `C.thrift`, `A`'s includes should contain `[A, C, B]` in that order. Because of the off by one error, we end up with `[A, B, B]`. This is because `B.includes()[0]` is `B`. `C` is at `B.includes()[1]`.

In practice, this leads to missing definitions + duplicate definitions.

Reviewed By: pranavtbhat

Differential Revision: D65635941

fbshipit-source-id: 3ee4d741d512ef405d3f607d6cf3b5a1bcc04af1