Part1/N: Allow use of Thrift Compact protocol
Summary:
Raw thrift client defaults to Compact currently with only switching to Binary if transport type is Http.

However, we need the ability at caller level to be able to use Binary mode for services that are starting to use Curse serialization e.g. mysqld.

https://www.internalfb.com/intern/staticdocs/thrift/docs/features/serialization/cursor/

```
Only Binary protocol is supported. Clients must be configured properly to interact with opted-in methods. Requests using the default Compact protocol will be completed with a TApplicationException.
```

Using compact mode will lead to errors like this with such services:
```
ApplicationException(ApplicationException { message: "Single pass serialization only supports binary protocol.", type_: UnsupportedClientType })
```

Considering compact and binary are the only wire level serializations possible, I am just using the with_protocol provide value to flip a bool in the FFI (this seemed easier/more-readable than conversions from fbthrift::thrift_protocol::ProtocolID.

(See next diff in stack for usage in a mysqld specific scenario which was previously broken)

What this means for Mysql Infra is that our ability to monitor sql over thrift status of our fleet is broken as a result of the single pass error (well not yet, but when latest mysqld rolls out to fleet - which is in progress atm). Considering all of Tao hits UDB over thrift, this would be a huge chink in our monitoring of the fleet and SLO calc.

Reviewed By: fadimounir, Imxset21

Differential Revision: D65855410

fbshipit-source-id: f713dce5fe3e90f71acece5450fa8ac893d268fb