Fix interaction name generation
Summary:
Interaction factory methods were setting the *service* name not the interaction name in the interaction_name field. Turns out we have a `{{function:interaction_name}}` key that seems to be undocumented, so we can use that (it's used in other parts of Rust thrift)

UPDATE: removed the following logic since it may require some upstream changes to do right. this only fixes the mustache generation for now
~~This additionally removes the usage of `getRequestContext` which is apparently a defunct hack. This requires interactions to hard-specify the protocol which the processor can be generated with, and we're picking binary (sorry everyone else)~~

Reviewed By: dtolnay

Differential Revision: D65903370

fbshipit-source-id: 4f97f57dcdfac1641628b202a3e1bab10af48a59