Allow root scope to be popped in eval_context
Summary:
This is a no-op in practice but noticed it in passing.

The current implementation doesn't allow the root context object to be popped. This is not necessary. In fact, with support for functions in the future, we may want to be able to clear all scopes (except global) to offer some degree of "isolation". Functions should not execute in the lexical scope of the caller.

Reviewed By: yoney

Differential Revision: D66141210

fbshipit-source-id: 7bee967d01321427aa82dde8f75c760cd92c4427