Fix issue with dunder enum arms in constants
Summary:
Currently, thrift-python blows up on `import` if an `enum` has an dunder `__` prefix arm, and that arm is used in a `const`. This is because python mangles class variables `__MyName` to `_MyClass__MyName`.

The smaller-scope way to handle this is to just correct the enum values for `const` in `.thrift_types` module.

V5: # / buildall

Reviewed By: Filip-F

Differential Revision: D66255826

fbshipit-source-id: 7e78120ff97bebd9e2bb256ae6c3e7ae3d5d6012