Use reserved identifiers to prevent conflict with symbols in the IDL file
Summary:
# What?
If any type has a field named `property`, then the use of the built-in `property` will conflict and fail. For an example of such a field that caused a failure, see `TestPropertyAsField` in `thrift/test/thrift-python/struct_test.thrift`.

# Fix
Use a reserved identifier alias for the built-in `property`.

Build fixtures with
```
buck2 run fbcode//thrift/compiler/test:build_fixtures
```

Reviewed By: yoney

Differential Revision: D65553769

fbshipit-source-id: d96b832bd715ad43bdda754246f06d8465ec7320