Resubmit: Refactor Compression Functions to be Methods on Manager
Summary:
D65494301 collided in landing with another diff and was reverted. This diff
is a resubmission without any other changes.

Depends on D65494301.

This is a no-op refactor that moves the lowest level methods that invoke
compresssion from being free functions to instead being methods on a class;
the `CompressionManager`.

Meta:

For the moment:

1. This new class is totally stateless.

2. This class is instantiated every place we invoke these functions, just for
   the lifetime of making that call. (This is fine, because of the above.)

This diff is intended to serve as a base for two workstreams:

1. The refactoring that guanqing is doing. The goal of that refactor is now
   concrete: we want to instantiate this `CompressionManager` once per
   connection and then use that single instance in all of these places,
   rather than instantiating temporaries in-line.

2. Moving the compression state / config into this class, cleaning it up from
   being stored elsewhere, and evolving the API this class exposes to be
   higher level.

Reviewed By: cevans87

Differential Revision: D66367019

fbshipit-source-id: 71f139d34fa52bd65a58be10e97900e4b2407377