Implement tests for special cases of test failures.
Summary:
To test the fixes for special cases of failures, add tests for them.

This diff contains the first of many. It moves a test case from `thrift/test/thrift-python`, which is meant for tests that serve as documentation of API contract. The special cases tests are sort of the opposite of that. They cover specific uncommon (anti-)patterns that we do not want to encourage.

Reviewed By: yoney

Differential Revision: D66433222

fbshipit-source-id: 0e9b7674a17fcb06350d8de79bd2219b4cc09117