skip on compression
Summary:
Updates the how hashing is skipped when compression is enabled. Looks at the rpc metadata struct to determine if the data is compressed or not be hashing. Hashing can be skipped when compressing because compression provides its own hashing. The checksumming will only run on uncompressed data.

Also improves the error message to match what the existing crc32 exception returns so this has the correct error codes.

Reviewed By: cevans87, avalonalex

Differential Revision: D66292663

fbshipit-source-id: e3d646b21f64e0df031a481dae94b0b065aee925