Fix: Type checks fail if a typename is the same as the field name.
Summary:
# What?
With type checks enabled for generated code, `special_cases.thrift` results in type errors like
```
Undefined or invalid type [11]: Annotation `TestFieldNameSameAsTypeName.TestStruct` is not defined as a type.
```

# Why?
Whenever the type checker encountered a found with a certain name, then that name is bound as a function. If that function name matches a type name, then that type name is no longer considered a type. With abstract types, fields are implemented as properties, which are functions.

# The Fix
This problem only affects the type names from the current module. Qualify the type names with a reference to the current module.

Reviewed By: ahilger

Differential Revision: D66664174

fbshipit-source-id: 2c976f9a61b5806b6be317c237145624b02d1d4a