ProtocolBench Build Rules (#626)
Summary:
This PR adds CMake build rules to build ProtocolBench. This will generate a `ProtocolBench` executable inside `bin/` in fbthrift's install directory.

This required a change to the ThriftLibrary, as generating sources from multiple thrift files with the same name (in different) directories was no supported. Previously, this led to multiple targets with identical name in CMake. By overloading the `thrift_generate` macro with an variant that takes a separate target name as an input, we add support for this while maintaining backwards compatibility.

Pull Request resolved: https://github.com/facebook/fbthrift/pull/626

Reviewed By: robertroeser

Differential Revision: D65667549

Pulled By: vitaut

fbshipit-source-id: d15957311ec9d3b54aba2926b298125c09efb619