Resolve circular references in user code
Summary:
# What?
Tests failed due to circular references that prevented buck's ability to build python binaries.

# Why?
The use of the "self-import" to set `_fbthrift_current_module` in the various types modules was the trigger. While the root cause might be elsewhere, I didn't see a strong enough reason to determine if we could make the `_fbthrift_current_module` solution work when there was previously proved workable solution available.

# Solution?
The original problem to solve was {D66664174}.

Use a reserved identifier as an alias for the types in the modules and use this alias as the type annotation rather than the original type. This prevents the problem that user functions, which, when named the same as a type, redefine the type as a function, and render that name unusable as a type annotation.

(Side Note: This was a change in a previous iteration before the use of the current module import, which we considered less attractive than the use of the fully qualified name with the current module at that time. Now we have different constraints, which changes our answer.)

# Reject options
- Obtain the current module via `sys.modules[__name__]`.
  - Rejected because this works at run-time. It does not work at type-check time.
- Use `__fbthrift_` as the alias prefix.
   - Rejected because this leads to an undefined symbol due to the way that Python mangles the name.
```
class TestStruct(...): ...
__fbthrift_TestStruct = TestStruct

class TestFieldNameSameAsTypeName(...):
    @_fbthrift_property
    def FieldNameSameAsTypeName(def) -> __fbthrift_TestStruct: ...
```
The type-checker sees `__fbthrift_TestStruct` as `_TestFieldNameSameAsTypeName_fbthrift_TestStruct`, which is not defined.

Reviewed By: prakashgayasen

Differential Revision: D66932282

fbshipit-source-id: dd143ecd351ff3172a5b9efbac33affa6c57471d