Add ODS Counter for logging received and successful requests for creating cases
Summary:
We are using thrift service for providing our case creation API from Care Platform to BSE team in fbcode.
They will call a python method, that calls our case creation method in www and all of that is implemented with thrift service.

In care_platform_client.py, we have an implementation of that thrift service that calls ServiceRouter, which routes requests to appropriate WWW service, which creates cases in Salesforce.

In this diff I have added an ODS counter(OBC, the library used here, is an implementation of it in python), which will log all received and successfully created cases in SF for tracking performance of the service in prod.

Differential Revision: D66883904

fbshipit-source-id: c9b59115cccc323974a0b58e742cacd7fb25be07