use type int for comparisons not enum
Summary:
It's inefficient to access `self.type.value`, specifically the second `.value` access, which is pure python `enum.Enum`. We can make a statically typed cython access on the internal data `int` instead.

This fixes a bug that occurs when unions with a field called `type` are compared. This fix is agnostic about what direction we want to take this API.

Reviewed By: yoney

Differential Revision: D67160589

fbshipit-source-id: 6d6471623ed11985e5bf0d334363cec89a2f44a9