SyntaxGraph API
Summary:
This diff introduces the `SyntaxGraph` API built on top of `schema.thrift` and exposes *most* of the useful information in an ergonomic C++ API while hiding implementation details such as `DefinitionKey`.

Type-safety and ergonomic are highest priority for this API — it should be *hard* to write incorrect code against this API. Some measures we've taken to encourage this:
- The `SyntaxGraph` class is the only object the user needs to keep alive. All "nodes" are returned by `const&` to hide memory management shenanigans from the user.
- All edges in the graph (usually `DefinitionKey` or URIs) are exposed as `const T&` or `const T*` instead.
- Things which cannot be null are `const T&` or `folly::not_null<const T*>`. Things which can be null are `const T* FOLLY_NULLABLE`.

This API is read-only — there is no mutation. The input can only be valid `schema.thrift` representations.

All `DefinitionKey` or `URI` references in the graph **must** exist or exceptions will be thrown. For bundled schema information, this will always be the case.

Reviewed By: pranavtbhat

Differential Revision: D66665902

fbshipit-source-id: 7c30e750aa200b82b6b4ae2c7e6a0986d1a022d1