Delete type information for services reflection
Summary:
With no uses of services reflection remaining in production, deleting the type info. I'll delete the actual functionality in another diff since just to be safe I'd prefer to first land the pyre deletion and then the functionality.

Also deleting the services reflection test since that functionality is supposed to be going away.

Reviewed By: ahilger

Differential Revision: D67328764

fbshipit-source-id: 53a4abb68c20ea7f928846f0ffac87d843dcfc49